+++ lib/src/currency_list_bottom_sheet.dart
@@ -7,7 +7,10 @@ void showCurrencyListBottomSheet({
   @required BuildContext context,
   @required ValueChanged<Currency> onSelect,
   List<String> currencyFilter,
+  String searchHint,
   bool showFlag = true,
+  bool showCurrencyName = true,
+  bool showCurrencyCode = true,
 }) {
   assert(context != null);
   assert(onSelect != null);
@@ -15,7 +18,15 @@ void showCurrencyListBottomSheet({
     context: context,
     isScrollControlled: true,
     backgroundColor: Colors.transparent,
-    builder: (_) => _builder(context, onSelect, currencyFilter, showFlag),
+    builder: (_) => _builder(
+      context,
+      onSelect,
+      currencyFilter,
+      searchHint,
+      showFlag,
+      showCurrencyName,
+      showCurrencyCode,
+    ),
   );
 }
 
@@ -23,7 +34,10 @@ Widget _builder(
   BuildContext context,
   ValueChanged<Currency> onSelect,
   List<String> currencyFilter,
+  String searchHint,
   bool showFlag,
+  bool showCurrencyName,
+  bool showCurrencyCode,
 ) {
   final device = MediaQuery.of(context).size.height;
   final statusBarHeight = MediaQuery.of(context).padding.top;
@@ -49,7 +63,10 @@ Widget _builder(
     ),
     child: CurrencyListView(
       onSelect: onSelect,
+      searchHint: searchHint,
       showFlag: showFlag,
+      showCurrencyName: showCurrencyName,
+      showCurrencyCode: showCurrencyCode,
       currencyFilter: currencyFilter,
     ),
   );